updated exception

+++ src/main/java/com/example/Inventory/Management/Service/Impl/CategoryServiceImpl.java
@@ -1,9 +1,9 @@
 package com.example.Inventory.Management.Service.Impl;
 
 import com.example.Inventory.Management.Entity.Category;
+import com.example.Inventory.Management.Exception.CategoryNotFoundException;
 import com.example.Inventory.Management.Repository.CategoryRepository;
 import com.example.Inventory.Management.Service.CategoryService;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -23,7 +23,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public Category getCategoryById(Long id) {
         return categoryRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Category not found"));
+                .orElseThrow(() -> new CategoryNotFoundException(id));
     }
 
     @Override
@@ -33,12 +33,17 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        Category category = getCategoryById(id);
+        
+        if (!category.getProducts().isEmpty()) {
+            throw new RuntimeException("Cannot delete category that has associated products");
+        }
+        
         categoryRepository.deleteById(id);
     }
 
     @Override
     public Category updateCategory(Long id, Category category) {
-        
         Category existingCategory = getCategoryById(id);
         existingCategory.setName(category.getName());
         existingCategory.setDescription(category.getDescription());
